Created Enough UI to Start Coding and Implemented Time
Made and implemented assets for everything on the top bar, everything except deity portraits for the deity panels, tabs, the main screen frame, and the job assignment screen.

Made GameManager script to handle overview of game. Implemented in-game time passing in the GameManager using delta time and an onDayPassed event called through the static EventManager class which takes the current day, month, and year as parameters.

Made a script for the tabs to switch between selected/unselected sprites and shift text up/down appropriately on hover/unhover and switch screens on click.

Made a script for the clock to update the time text on day passed and a script for the UI slider that sets the time scale equal to the slider int value converted to the appropriate multiplier.

Started making framework for Deity class.