Implemented Game Over and UI Scaling on Mouse Over
Player now loses when any Deity's happiness hits 0. The tabs change to only have 2 large tabs, "Game Over" and "Stats", and the main screen switches to the new Game Over screen which just shows "Game Over" text and replay/main menu buttons. The game speed slider is set to 0 (pausing the game) and set un-interactable.

Made a new ScaleOnMouseOver script that scales an object to a specified scale with a specified duration and applied it to the game over buttons, worker assignment buttons, and game speed slider handle.